+++ src/storm/simulator/PrismProgramSimulator.h
@@ -39,6 +39,7 @@ namespace storm {
             /**
              *
              * @return A list of choices that encode the possibilities in the current state.
+             * @note successor states are encoded using state indices that will potentially be invalidated as soon as the internal state of the simulator changes
              */
             std::vector<generator::Choice<ValueType, uint32_t>> const& getChoices() const;