+++ src/storm/storage/expressions/ExpressionStringFormat.h
@@ -1,9 +1,0 @@
-#pragma once
-
-namespace storm {
-    namespace expressions {
-        struct ExpressionStringFormat {
-            bool alternativeIff = false;
-        };
-    }
-}